Added the geoJson
The second iteration of the geojson file. This one was data scraped and
actually has the data. V-Dem is a lovely data source, but it did not
contain latitude and longitude for the countries, nor the polygon
coordinates. So, I had to merge a data-scraped CSV file with a hand-made
geojson file to make the next iteration.