I made a map :)
On a more serious note, an Interactive Choropleth Map was created based
on the liberal democracy index. The main error I notice is that the
color change is a bit weird when switching the years, but it's not that
big of a deal. The interactivity is there, inclding hovering over
countries and having multiple tabs.