Created a custom geoJson
My magnum opus. Pretty much I scraped the data from the V-Dem and
applied it to yearly columns, attempting to match a CSV filed that was
created from a custom geoJson. I then merged this CSV into a new geoJson
(because nothing could apparently process that the CSV had coordinates),
and then we now have a map with real data plus years. I would have loved
to do it as a slider, but such are the limitations of no filter.