Added the trends and populism tabs
Unfortunatly I forgot to keep up with saving images and skipped a
commit, but this covers now two major updates. Pretty much I added two
tabs, one encompasses the general trend in the decline of the liberal
democracy index, grouped by country and the "regime type". The other tab
is focused on two things, one is a regression analysis on
populism/anti-pluralism and democratic backsliding, and the other is the
rise of populism in recent years.

The UI specifically covers numerous changes, such as adding said tabs,
adding a new type of conditional panel, and a radio-type button.