Added two new tabs
Unfortunatly I forgot to keep up with saving images and skipped a
commit, but this covers now two major updates. Pretty much I added two
tabs, one encompasses the general trend in the decline of the liberal
democracy index, grouped by country and the "regime type". The other tab
is focused on two things, one is a regression analysis on
populism/anti-pluralism and democratic backsliding, and the other is the
rise of populism in recent years.

The server side was much more complicated than the UI side. To briefly
discuss what exactly was expanded upon compared to making the map,
reactive graphs were added, including having them be clickable, with one
displaying coordinates and the other displaying summary of data (a
renderprint).