Added polarization and inequality tabs
The polarization tab had to be broken down because of the regions,
resulting in a bloated amount of variables that essentially did the same
thing just to extend the plot.The regression plot itself, though, was
fairly simple. Clicking on each graph also displays the significance,
which is helpful. The Gini was harder, requiring two lines to exist at
once for each switch. This is why rbind was used. Otherwise it does the
same thing.